refactor: message lists query invalidation in custom hook for readability
also: cleaned out some dead code/translation keys
